code split: add display products component
